[RF][HS3][HF] Merge `testHSHistFactory` with `testHistFactory`
All HistFactory unit tests are merged in one file, such that the nice
test matrix in `testHistFactory` can also be used for the HS3
HistFactory tests.

In particular, the combined test is also covering the support for
non-uniform binning in HistFactory HS3.
