+++ roofit/hs3/test/CMakeLists.txt
@@ -1,3 +1,2 @@
 ROOT_ADD_GTEST(testRooFitHS3 testRooFitHS3.cxx LIBRARIES RooFitCore RooFit RooFitHS3)
 ROOT_ADD_GTEST(testHS3SimultaneousFit testHS3SimultaneousFit.cxx LIBRARIES RooFitCore RooFit RooFitHS3 RooStats)
-ROOT_ADD_GTEST(testHS3HistFactory testHS3HistFactory.cxx LIBRARIES RooFit RooFitHS3 HistFactory)
